+++ doc/whats_new.rst
@@ -17,6 +17,8 @@ Changelog
 
     - Add filter plotting functions :func:`mne.viz.plot_filter` and :func:`mne.viz.plot_ideal_filter` as well as filter creation function :func:`mne.filter.create_filter` by `Eric Larson`_
 
+    - Add TPS warping with spherical harmonic surface approximations in :class:`mne.transforms.SphericalHarmonicTPSWarp` by `Eric Larson`_
+
 BUG
 ~~~